Move configs to sub directories
